vision/requirements.txt
requests
numpy
pandas
tqdm
plotnine
scikit-misc
tensorflow==2.2.1
nibabel
matplotlib
pillow
batchgenerators
pydicom
SimpleITK
scikit-image
dicom2nifti
watchdog
pymongo
python-dotenv
python-magic
pynetdicom